OpenChat-main/chat_app/requirements.txt
flask==3.0.2
flask-socketio==5.3.6
python-socketio==5.11.1
python-engineio==4.9.1
pillow==10.2.0